+++ plugins/frontend/backstage-plugin-prometheus/src/api/index.ts
@@ -22,6 +22,7 @@ import {
 import { DateTime, Duration } from 'luxon';
 
 const DEFAULT_PROXY_PATH = '/prometheus/api';
+const SERVICE_NAME_HEADER = 'x-prometheus-service-name';
 
 export const prometheusApiRef = createApiRef<PrometheusApi>({
   id: 'plugin.prometheus.service',
@@ -44,6 +45,23 @@ export class PrometheusApi {
   private async getApiUrl({ serviceName }: { serviceName?: string }) {
     const proxyUrl = await this.discoveryApi.getBaseUrl('proxy');
     return `${proxyUrl}${this.getProxyPath({ serviceName })}`;
+  }
+
+  public getUiUrl({ serviceName }: { serviceName?: string }) {
+    if (Boolean(serviceName)) {
+      const instances = this.configApi.getOptionalConfigArray(
+        'prometheus.instances',
+      );
+      if (instances && instances?.length > 0) {
+        const instance = instances.find(
+          value => value.getString('name') === serviceName,
+        );
+        if (Boolean(instance)) {
+          return instance?.getOptionalString('uiUrl');
+        }
+      }
+    }
+    return this.configApi.getOptionalString('prometheus.uiUrl');
   }
 
   private getProxyPath({ serviceName }: { serviceName?: string }) {
@@ -87,6 +105,11 @@ export class PrometheusApi {
     const start = DateTime.now().minus(Duration.fromObject(range)).toSeconds();
     const response = await fetch(
       `${apiUrl}/query_range?query=${query}&start=${start}&end=${end}&step=${step}`,
+      {
+        headers: {
+          [SERVICE_NAME_HEADER]: serviceName || '',
+        },
+      },
     );
     if (!response.ok) {
       throw new Error(
@@ -98,7 +121,11 @@ export class PrometheusApi {
 
   async getAlerts({ serviceName }: { serviceName?: string }) {
     const apiUrl = await this.getApiUrl({ serviceName });
-    const response = await fetch(`${apiUrl}/rules?type=alert`);
+    const response = await fetch(`${apiUrl}/rules?type=alert`, {
+      headers: {
+        [SERVICE_NAME_HEADER]: serviceName || '',
+      },
+    });
 
     if (!response.ok) {
       throw new Error(