+++ plugins/frontend/backstage-plugin-cloudsmith/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-cloudsmith
 
+## 1.0.14
+
+### Patch Changes
+
+- c6b0af08: Bump plugins version to backstage version 1.18.3
+
 ## 1.0.13
 
 ### Patch Changes