+++ plugins/frontend/backstage-plugin-jira/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-jira
 
+## 2.4.11
+
+### Patch Changes
+
+- c6b0af08: Bump plugins version to backstage version 1.18.3
+
 ## 2.4.10
 
 ### Patch Changes