+++ plugins/home/backstage-plugin-home-markdown/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 1.1.0
 
+## 2.5.15
+
+### Patch Changes
+
+- c6b0af08: Bump plugins version to backstage version 1.18.3
+
 ## 2.5.14
 
 ### Patch Changes