+++ plugins/scaffolder-actions/scaffolder-backend-argocd/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @roadiehq/scaffolder-backend-argocd
 
+## 1.1.17
+
+### Patch Changes
+
+- c6b0af08: Bump plugins version to backstage version 1.18.3
+- Updated dependencies [c6b0af08]
+  - @roadiehq/backstage-plugin-argo-cd-backend@2.11.3
+
+## 1.1.16
+
+### Patch Changes
+
+- 2fbcbee1: Bump to Backstage version 1.17.5
+
+## 1.1.15
+
+### Patch Changes
+
+- af8fc26a: fix: return item when no history on argo app
+- Updated dependencies [56c40ed5]
+  - @roadiehq/backstage-plugin-argo-cd-backend@2.11.2
+
 ## 1.1.14
 
 ### Patch Changes