+++ plugins/scaffolder-actions/scaffolder-backend-module-http-request/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/scaffolder-backend-module-http-request
 
+## 4.0.13
+
+### Patch Changes
+
+- c6b0af08: Bump plugins version to backstage version 1.18.3
+
+## 4.0.12
+
+### Patch Changes
+
+- 2fbcbee1: Bump to Backstage version 1.17.5
+
 ## 4.0.11
 
 ### Patch Changes